pyproject.toml
[tool.poetry]
name = "rag"
version = "0.1.0"
description = ""
authors = ["sid-thephysicskid <siddabhinav@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8.1,<3.9.7 || >3.9.7,<3.13"
pinecone-client = {extras = ["grpc"], version = "^3.2.1"}
streamlit = "^1.32.2"
langchain = "^0.1.13"
openai = "^1.14.3"
tiktoken = "^0.6.0"
langchain-pinecone = "^0.0.3"
langchain-openai = "^0.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
